requirements/install.txt
absl-py # cli
appdirs
cloudpickle
cython
devtools
h5py
jsonargparse[signatures]
lightning
ml_collections==1.1.0 # Config interface. Need exact version since we overwrite internal functions
numpy>=1.21.0,<2.0.0
opencv-python
pandas
pillow
plyfile
pycocotools
pydantic<2.0.0
setuptools
tensorboard
termcolor
terminaltables
timm>=0.6.0
torch>=2.0.0
torchvision>=0.15.1
tqdm
utm
wheel
scipy
scipy-stubs

# packages for datasets
bdd100k
scalabel
nuscenes-devkit
